+++ retdec/decompiler.py
@@ -29,6 +29,9 @@ class Decompiler(Service):
         :type target_language: str
         :param decomp_var_names: Naming style for variables.
         :type decomp_var_names: str
+        :param decomp_optimizations: Type of optimizations performed by the
+            decompiler.
+        :type decomp_optimizations: str
         :param architecture: Architecture. The precise meaning depends on the
             used `mode`.
         :type architecture: str
@@ -83,6 +86,7 @@ class Decompiler(Service):
         }
         self._add_param_when_given('target_language', params, kwargs)
         self._add_param_when_given('decomp_var_names', params, kwargs)
+        self._add_param_when_given('decomp_optimizations', params, kwargs)
         self._add_param_when_given('architecture', params, kwargs)
         self._add_param_when_given('file_format', params, kwargs)
         self._add_param_when_given('comp_compiler', params, kwargs)