+++ tests/decompiler_tests.py
@@ -142,6 +142,15 @@ class DecompilerStartDecompilationTests(BaseServiceTests):
             params=AnyParamsWith(decomp_var_names='simple')
         )
 
+    def test_decomp_optimizations_is_set_to_correct_value_when_given(self):
+        self.start_decompilation_with_any_input_file(
+            decomp_optimizations='none'
+        )
+
+        self.assert_post_request_was_sent_with(
+            params=AnyParamsWith(decomp_optimizations='none')
+        )
+
     def test_architecture_is_set_to_correct_value_when_given(self):
         self.start_decompilation_with_any_input_file(
             architecture='arm'